+++ backend/sokdak/src/main/java/com/wooteco/sokdak/post/service/PostService.java
@@ -20,8 +20,10 @@ import com.wooteco.sokdak.post.dto.PostUpdateRequest;
 import com.wooteco.sokdak.post.dto.PostsResponse;
 import com.wooteco.sokdak.post.exception.PostNotFoundException;
 import com.wooteco.sokdak.post.repository.PostRepository;
+import com.wooteco.sokdak.profile.dto.MyPostsResponse;
 import java.util.Collections;
 import java.util.List;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
@@ -86,12 +88,20 @@ public class PostService {
         boolean liked = likeRepository.existsByMemberIdAndPostId(authInfo.getId(), postId);
         Hashtags hashtags = hashtagService.findHashtagsByPostId(postId);
 
-        return PostDetailResponse.of(foundPost, postBoards.get(0), liked, foundPost.isAuthenticated(authInfo.getId()), hashtags);
+        return PostDetailResponse.of(foundPost, postBoards.get(0), liked, foundPost.isAuthenticated(authInfo.getId()),
+                hashtags);
     }
 
     public PostsResponse findPostsByBoard(Long boardId, Pageable pageable) {
         Slice<PostBoard> postBoards = postBoardRepository.findPostBoardsByBoardId(boardId, pageable);
         return PostsResponse.ofPostBoardSlice(postBoards);
+    }
+
+    public MyPostsResponse findMyPosts(Pageable pageable, AuthInfo authInfo) {
+        Member member = memberRepository.findById(authInfo.getId())
+                .orElseThrow(MemberNotFoundException::new);
+        Page<Post> posts = postRepository.findPostsByMemberOrderByCreatedAtDesc(pageable, member);
+        return MyPostsResponse.ofPosts(posts.getContent(), posts.getTotalPages());
     }
 
     @Transactional
