+++ backend/sokdak/src/test/java/com/wooteco/sokdak/util/ControllerTest.java
@@ -22,8 +22,6 @@ import com.wooteco.sokdak.notification.controller.NotificationController;
 import com.wooteco.sokdak.notification.service.NotificationService;
 import com.wooteco.sokdak.post.controller.PostController;
 import com.wooteco.sokdak.post.service.PostService;
-import com.wooteco.sokdak.profile.controller.ProfileController;
-import com.wooteco.sokdak.profile.service.ProfileService;
 import com.wooteco.sokdak.report.controller.CommentReportController;
 import com.wooteco.sokdak.report.controller.PostReportController;
 import com.wooteco.sokdak.report.service.CommentReportService;
@@ -53,8 +51,7 @@ import org.springframework.web.context.WebApplicationContext;
         BoardController.class,
         HashtagController.class,
         NotificationController.class,
-        HashtagController.class,
-        ProfileController.class
+        HashtagController.class
 })
 @ExtendWith(RestDocumentationExtension.class)
 public class ControllerTest {
@@ -108,9 +105,6 @@ public class ControllerTest {
 
     @MockBean
     protected RefreshTokenService refreshTokenService;
-
-    @MockBean
-    protected ProfileService profileService;
 
     @BeforeEach
     void setRestDocs(WebApplicationContext webApplicationContext,
